+++ packages/enraiged/src/Users/Services/Support/UserProfileAttributes.php
@@ -29,23 +29,23 @@ class UserProfileAttributes extends AttributeHandler
             unset($this->attributes['name']);
         }
 
-        //  remove the password attribute if not in the provided collection
+        //  remove the password attribute if the value is null
         if (key_exists('password', $this->attributes) && is_null($this->attributes['password'])) {
             unset($this->attributes['password']);
         }
 
+        //  ensure a default password does exist, if none provided
+        if (!key_exists('password', $this->attributes)) {
+            $this->attributes['password'] = config('enraiged.auth.insecure_password');
+        }
+
+        //  attempt to set a role name if none exists
+        if (!key_exists('role_id', $this->attributes) && !key_exists('role', $this->attributes)) {
+            $this->attributes['role'] = config('enraiged.auth.force_default_role');
+        }
+
         //  determine the id of a role identified by its name
-        if (key_exists('role', $this->attributes)) {
-            /*
-            $role = preg_match('/\d+/', $this->attributes['role'])
-                ? Role::find($this->attributes['role'])
-                : Role::where('name', $this->attributes['role'])->first();
-
-            if ($role->exists) {
-                $this->attributes['role_id'] = $role->id;
-            }
-            */
-
+        if (key_exists('role', $this->attributes) && !is_null($this->attributes['role'])) {
             $this->attributes['role_id'] = Role::where('name', $this->attributes['role'])
                 ->first()
                 ->id;